refactor: restructure test suite 17→13 files, feature-focused organization
- Create test_features.py: consolidates v0.4.0 + v0.5.0 feature tests (seed,
  rank-norm, schema, export-validation, canonical naming, provenance, trial-db,
  neutralizer, masked aux targets)
- Create test_autoresearch.py: mutations, state, and one-trial integration test
- Create test_models.py: LightGBM/CatBoost smoke tests, model spec normalization,
  EnsembleOptimizer, SyntheticDataAugmenter
- Expand test_pipeline.py: absorbs stacking, internal-val-split, save/load,
  era-pca, NaN rows, numerai-predict, multi-target, and e2e experiment tests
- Expand test_hpo_pipeline.py: absorbs LightGBM round threading and XGBoost
  default round tests
- Delete test_review_fixes.py, test_review_remediation.py, test_v040_features.py,
  test_v050_features.py (all content redistributed; process-named files removed)
